pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agar"
version = "0.1.0"
description = "AGAR colony detection project (Hydra + Fabric + MLflow + FiftyOne-ready)"
requires-python = ">=3.11"
dependencies = [
  "torchmetrics>=1.4.0",
  "hydra-core==1.3.2",
  "omegaconf>=2.3.0",
  "lightning-fabric==2.6.0",
  "tensorboard>=2.16.0",
  "albumentations>=1.4.0",
  "opencv-python>=4.9.0",
  "tqdm>=4.66.0",
  "rich>=13.7.0",
  "mlflow>=3.8.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.6.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
